requirements.txt
pyyaml
boto3
librosa
Pillow
torch
torchvision
numpy
psutil
matplotlib
scikit-learn
ffmpeg-python
pandas
django
gunicorn

